Initial script to generate Core manifests
This script will iterate over all of the core snaps
installed on a system, and generate manifest files
for each which list the debian packages staged by
each. If will then iterate over these manifest files
adding stage packages from any snaps that declare
the corresponding core snap as a base.

A dedicated manifest will also be generated for snapd,
which itself is based on 16.04 LTS.

Signed-off-by: Tony Espy <espy@canonical.com>