[WIP] report: use xpath to grab CVE priorities
This is a proof-of-concept to determine if using xpath/xmlstarlet
to grab CVE priorities for result=true results in an OSCAP OVAL
report file. The performance penalty for this approach will most
likely make it unusable. The code uses a single xpath expression
to grab all the 'true' CVEs, but then needs to use a single expr
to grab each priority from the <definitions>.

Signed-off-by: Tony Espy <espy@canonical.com>